[MO] Implement TensorFlow 2 While and Keras RNN support in MO (#3573)
* [MO] Implement TensorFlow 2 While support in MO


* Add extractors for both While and StatelessWhile and do minor changes


* Improve update_body_graph function and manage graph names properly


* Fix a map for original name of parameters from body and cond


* Implement draft version of support of TF2 Keras RNN


* Implement Keras LSTM and GRU support in MO


* Improve code for Keras RNN support


* Finalize implementation of TF2 Keras RNN support in MO


* Apply the first part of the comments after review #1


* Avoid use of explicit values of port indices in the transformation


* Finalize code after the first-round review


* Apply comments after the second-round review

+++ model-optimizer/extensions/ops/loop.py
@@ -1,5 +1,5 @@
 """
- Copyright (C) 2017-2020 Intel Corporation
+ Copyright (C) 2017-2021 Intel Corporation
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -94,6 +94,9 @@ class Loop(TensorIterator):
                 loop_port_idx = record['external_port_id']
                 if loop_port_idx != -1:
                     input_shape = loop_node.in_port(loop_port_idx).get_connection().get_source().data.get_shape()
+                slice_axis = record['axis']
+                if slice_axis is not None:
+                    input_shape[slice_axis] = 1
                 body_node.shape = input_shape
                 log.debug('Updated shape for the body node with internal_id "{}" with value {}'
                           ''.format(record['internal_layer_id'], body_node.shape))
@@ -155,6 +158,8 @@ class Loop(TensorIterator):
         num_iterations = loop_node.in_port(0).data.get_value()
         if num_iterations is not None:
             num_iterations = num_iterations.item(0)
+            if num_iterations < 0:
+                return None
         return num_iterations
 
     @staticmethod
@@ -317,8 +322,56 @@ class Loop(TensorIterator):
             if record[attr] == original_value:
                 record[attr] = new_value
                 matched += 1
-        assert matched == 1, 'More than one record in the portmap for attr "{}" wil original value "{}"' \
+        assert matched == 1, 'More than one record in the portmap for attr "{}" with original value "{}"' \
                              ''.format(attr, original_value)
+
+    @staticmethod
+    def update_port_map_value_ext(port_map: dict, layer_id_attr: str, layer_id_value: int,
+                                   updated_attr: str, new_attr_value: int):
+        """
+        Updates a value of requested attribute for a certain layer id in a port map
+        :param port_map: a map of external ports to internal layer ids
+        :param layer_id_attr: layer id attribute for which to update attribute
+        :param layer_id_value: layer id value for which to update attribute
+        :param updated_attr: a name of attribute which to update
+        :param new_attr_value: new value of attribute
+        """
+        matched = 0
+        for record in port_map:
+            if record.get(layer_id_attr) == layer_id_value:
+                record[updated_attr] = new_attr_value
+                matched += 1
+        assert matched == 1, 'More than one record in the portmap for attr "{}" with original value "{}"' \
+                             ''.format(layer_id_attr, layer_id_value)
+
+    @staticmethod
+    def back_edge_exists(back_edges_map: dict, from_layer: int, to_layer: int):
+        """
+        Checks if a back edge exists in the back_edges_map connecting specific nodes
+        :param back_edges_map: a map where to search for specified back edge
+        :param from_layer: id of Result node that belongs a back edge
+        :param to_layer: id of Parameter node that belongs a back edge
+        :return: True or False
+        """
+        for back_edge in back_edges_map:
+            if back_edge['from_layer'] == from_layer and back_edge['to_layer'] == to_layer:
+                return True
+        return False
+
+    @staticmethod
+    def inter_edge_exists(port_map: dict, external_port_id: int, internal_layer_id: int):
+        """
+        Check if inter-graph edge (i.e. an edge between the main graph and body graph) exists
+        :param port_map: a port map where to search for inter-graph edge
+        :param external_port_id: port index from/to which edge goes
+        :param internal_layer_id: layer id from/to which edge goes
+        :return: True or False
+        """
+        for i_port in port_map:
+            if i_port['external_port_id'] == external_port_id and \
+                    i_port['internal_layer_id'] == internal_layer_id:
+                return True
+        return False
 
     @staticmethod
     def re_numerate_input_ports(loop_node: Node):
@@ -372,7 +425,8 @@ class Loop(TensorIterator):
                     new_port_id += 1
 
             for port_idx_to_remove in reversed(range(new_port_id, max_port_id + 1)):
-                loop_node.delete_output_port(port_idx_to_remove)
+                if port_idx_to_remove in loop_node.out_ports().keys():
+                    loop_node.delete_output_port(port_idx_to_remove)
 
     @staticmethod
     def remove_unused_ops_from_port_map(loop_node: Node, port_map: dict, port_map_attr: str, dir: [None, str] = None):
